Split path.rb from mkrunnable.rb
