+++ tool/mkrunnable.rb
@@ -6,6 +6,7 @@
 
 require './rbconfig'
 require 'fileutils'
+require_relative 'lib/path'
 
 case ARGV[0]
 when "-n"
@@ -18,93 +19,7 @@ else
   include FileUtils
 end
 
-module Mswin
-  def ln_safe(src, dest, *opt)
-    cmd = ["mklink", dest.tr("/", "\\"), src.tr("/", "\\")]
-    cmd[1, 0] = opt
-    return if system("cmd", "/c", *cmd)
-    # TODO: use RUNAS or something
-    puts cmd.join(" ")
-  end
-
-  def ln_dir_safe(src, dest)
-    ln_safe(src, dest, "/d")
-  end
-end
-
-def clean_link(src, dest)
-  begin
-    link = File.readlink(dest)
-  rescue
-  else
-    return if link == src
-    File.unlink(dest)
-  end
-  yield src, dest
-end
-
-def ln_safe(src, dest)
-  ln_sf(src, dest)
-rescue Errno::ENOENT
-  # Windows disallows to create broken symboic links, probably because
-  # it is a kind of reparse points.
-  raise if File.exist?(src)
-end
-
-alias ln_dir_safe ln_safe
-
-case RUBY_PLATFORM
-when /linux|darwin|solaris/
-  def ln_exe(src, dest)
-    ln(src, dest, force: true)
-  end
-else
-  alias ln_exe ln_safe
-end
-
-if !File.respond_to?(:symlink) && /mingw|mswin/ =~ (CROSS_COMPILING || RUBY_PLATFORM)
-  extend Mswin
-end
-
-def clean_path(path)
-  path = "#{path}/".gsub(/(\A|\/)(?:\.\/)+/, '\1').tr_s('/', '/')
-  nil while path.sub!(/[^\/]+\/\.\.\//, '')
-  path
-end
-
-def relative_path_from(path, base)
-  path = clean_path(path)
-  base = clean_path(base)
-  path, base = [path, base].map{|s|s.split("/")}
-  until path.empty? or base.empty? or path[0] != base[0]
-    path.shift
-    base.shift
-  end
-  path, base = [path, base].map{|s|s.join("/")}
-  if /(\A|\/)\.\.\// =~ base
-    File.expand_path(path)
-  else
-    base.gsub!(/[^\/]+/, '..')
-    File.join(base, path)
-  end
-end
-
-def ln_relative(src, dest, executable = false)
-  return if File.identical?(src, dest)
-  parent = File.dirname(dest)
-  File.directory?(parent) or mkdir_p(parent)
-  if executable
-    return (ln_exe(src, dest) if File.exist?(src))
-  end
-  clean_link(relative_path_from(src, parent), dest) {|s, d| ln_safe(s, d)}
-end
-
-def ln_dir_relative(src, dest)
-  return if File.identical?(src, dest)
-  parent = File.dirname(dest)
-  File.directory?(parent) or mkdir_p(parent)
-  clean_link(relative_path_from(src, parent), dest) {|s, d| ln_dir_safe(s, d)}
-end
+include Path
 
 config = RbConfig::MAKEFILE_CONFIG.merge("prefix" => ".", "exec_prefix" => ".")
 config.each_value {|s| RbConfig.expand(s, config)}
